Fixed a crash with finding what files a proxy type might have come from after a file rename + delete. --force-push-
Summary: There's two important maps on the resource system: the text resources and the actual resources. There was a chance that a resource could be removed from the global map but not the text resource map during deletion which would cause FindProxiedTypeOrigin to crash since it didnt' check for null after looking up a resource by id. I wrapped this with a null check but additionally fixed the root issue. The TextResources map was of file path to resource id. When a resource was renamed it didn't remove the old file path from the map. If you then deleted this renamed resource then there was a path mapping to a removed resource id.

Differential Revision: https://www.zeroengine.io/D166