Fixed object view create menu sometimes not properly creating child objects.-
Summary:
Moved setting the command context for the creation commands to mouse entering the create menu.-
Changed clearing the command context to being on focus lost for the submenu over the menu closed event on the parent menu which sometimes resulted in not properly creating child objects. Fixes T1052.-

Test Plan: go the object view and right click an object and go to the create sub menu. Hold left click on an item to create as a child and let go. It should create the child object properly.

Reviewers: joshua.claeys

Reviewed By: joshua.claeys

Subscribers: dane.curbow

Maniphest Tasks: T1052

Differential Revision: https://www.zeroengine.io/D217