When changing certain properties on UiWidget, not all side effect properties were being properly marked as modified.
Summary: Property side effects are not recursive... They need to be, but that's a larger issue. This is a temporary fix.

Test Plan: Create an Archetype with a UiWidget. Have widget be fixed on the Y with a layout and a few children. Make it auto sized on the Y (SizePolicyY = Auto), and it should properly mark UiWidget.Size as modified as well as Area.Size.

Reviewers: #zero_engine_developers, joshua.davis

Reviewed By: #zero_engine_developers, joshua.davis

Subscribers: joshua.claeys

Differential Revision: https://www.zeroengine.io/D326