UiHighlight now respects the HandledEvent flag on mouse events.+
Summary: I ran into this bug when trying to create a title bar for a window with a close button on the title bar. The title bar would highlight when the mouse went over it, but was still responding to events when the mouse was over the close button.

Test Plan: Make a widget with the UiHighlight component. Add another widget as a child that responds to and handles the LeftMouseDown / Up events.

Reviewers: #zero_engine_developers, andrea.ellinger

Reviewed By: #zero_engine_developers, andrea.ellinger

Subscribers: joshua.claeys

Differential Revision: https://www.zeroengine.io/D333