Shift Select Fix
Summary: Broke archetype/root select out into a helper function (SmartSelect) for use with single select and multi-select and made the functions static for use within ZilchScript. Changed ArchetypeSelect function to static as it is used within the SmartSelect function. Placed the logic for selecting an object that is part of a hierarchy and within its parents AABB after the call to SmartSelect to remove the need for SmartSelect to take in a RayCastResultList.

Test Plan: Open a project and click on objects within an archetype, prefereably one composed of other archetypes.

Reviewers: joshua.claeys

Reviewed By: joshua.claeys

Subscribers: dane.curbow

Differential Revision: https://www.zeroengine.io/D350