CogPath Property Grid Add Component
Summary: Changed check for add button from enumerating all types to just types that can be added to the instance the property grid is generating a display for. Added a can add component check to CogPathMetaComposition that returns false as the base MetaComposition assumes the object it is present on can handle adding components in general before doing dependency checks. Fixes T1351.

Test Plan: Add a CogPath as a property to an object and view it in the property grid. There will be no Add Component at the bottom of the CogPaths expanded view.

Reviewers: joshua.claeys

Reviewed By: joshua.claeys

Subscribers: dane.curbow

Maniphest Tasks: T1351

Differential Revision: https://www.zeroengine.io/D369