Made zilch templates now appear in documentation. Fixes T507.-
Summary:
Changed LoadFromMeta to create instances of templates in libraries, then replace the placeholder types used
to create an instantiation of that template with the template parameter names so when written to a documentation file
it will show the template parameter name (i.e. HashMap[Key, Value] instead of HashMap[Wrapper,UTF8Encoding]). The types
that were used to instantiate were Wrapper and UTF8Encoding because after checking all the templates, these types are not
used so are safe to just replace with the template param names after the template is instantiated. An actual Zilch type
was needed in order to be able to create an instance, and this route was chosen as to not have to add a fake type to the engine.

Test Plan: N/A

Reviewers: joshua.davis

Reviewed By: joshua.davis

Subscribers: joshua.davis, joshua.shlemmer

Maniphest Tasks: T507, T1150

Differential Revision: https://dev.zeroengine.io/D396