Hopefully fixed the launcher (a bit of a hack) to not stay open by forcing the string pool to be released when a crash happens.-
Summary: The launcher crashed on occasion (not sure why it crashed), typically after installing a patch, and it would infinite loop on the spin lock of a string causing the launcher to forever stay as a background process. This is a quick hack fix to try to make this never happen.

Test Plan: I had to add a lot of code to force to test this. Come bug me on my machine to test.

Reviewers: trevor.sundberg

Subscribers: joshua.davis

Differential Revision: https://www.zeroengine.io/D422