Fixed how the launcher determined if the eula had been accepted to use the eula's hash instead of a timestamp.-
Summary: The launcher was using the timestamp of the eula file which was constantly getting updated everytime a new patch was made. This switches it to use the hash of the eula so if the text ever changes the eula will display again.

Test Plan: Verify the eula shows up once and never again unless you modify the file's contents.

Reviewers: #zero_engine_developers

Subscribers: joshua.davis

Differential Revision: https://www.zeroengine.io/D425