Attempting to get more information about a common crash when core resources can't be copied. Added more logging and more attempts to fix the problem.-
Summary: Crashes like https://crash.zeroengine.io/T3928. In these scenarios they couldn't copy a core file for some reason. In an attempt to get more information I added another function that prints where a path is missing if it is. The path should always exist by this point in time but might as well try and log more info. Additionally added an extra attempt to create missing directory parents just in case.

Test Plan: Put a breakpoint in BuildContent before the file copy. When you get it to hit (you might have to clear your cache first). Delete one of the parent directories of this path to make CopyFile fail. You'd have to do this twice to get the error messages to trip now.

Reviewers: #zero_engine_developers

Subscribers: joshua.davis

Differential Revision: https://www.zeroengine.io/D426