Collapse pivot option for importing models.
Summary:
Added the option to collapse pivots when importing meshes.
Added whether a node is a pivot to the hierarchy data for imported models. Removed commented out code.
When chosing to collapse pivots the models are collapsed properly. When collapsing pivots animated nodes are not collapsed and animation tracks are corrected.
Added comment to collapse pivot option to warn users of potential consequence to collapsing pivots on a model intending on using non-embedded animations.

Test Plan: n/a

Reviewers: nathan.carlson

Reviewed By: nathan.carlson

Subscribers: dane.curbow

Differential Revision: https://dev.zeroengine.io/D401