Removed raycast select through root object in archetype when the root is selected that results in weird behavior as the raycast doesn't differentiate between aabb and graphical selection.-
Summary: Fixes odd archetype select behavior, need to re-evalute whether the idea behind the behavior was valid to begin with.

Test Plan: Open a scene with an archetype and select the root and select it again when it has a child object behind it. It should select the root object.

Reviewers: nathan.carlson

Reviewed By: nathan.carlson

Subscribers: dane.curbow

Differential Revision: https://dev.zeroengine.io/D473