When multiselecting objects in the property grid and setting a property on a component or creating a new resource to set on a component change the value on all the components in the selection.
Summary: Fixes T1578.

Test Plan: Create a few objects in a level with a model component. Multi-select the objects and change the material on the model. Crarte a new material while multiselecting the objects with model present. When changing or creating materials it wil be set on all the model components for all currently selected cogs.

Reviewers: #zero_engine_developers, trevor.sundberg, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: joshua.claeys, trevor.sundberg, dane.curbow

Maniphest Tasks: T1578

Differential Revision: https://dev.zeroengine.io/D513