Fixed a crash when adding Components to the selection and recompiling scripts.+
Summary: Property grid was storing Handles to types that were being deleted when scripts compile.

Test Plan: Write an editor script to add multiple Zilch Components of the same type to the editor's selection. Modify scripts and it should no longer crash when doing any operation in the property grid.

Reviewers: #zero_engine_developers, andrea.ellinger

Reviewed By: #zero_engine_developers, andrea.ellinger

Subscribers: joshua.davis, joshua.claeys

Differential Revision: https://dev.zeroengine.io/D528