Fixed no-op restore locally removed component
Summary: Only show the restore component option in the property grid if the component can be restored instead of resulting in a no-op when a sibling component has been added (i.e BoxCollider, SphereCollider). Moved ContextMenu creation to within the branching cases so that it doesn't create an empty menu that we need to delete if these two cases are not hit. Fixes T1453.

Test Plan: Create an archetype from an object with a box collider. Remove the box collider. Add a sphere collider. Right click to restore the box collider. No menu.

Reviewers: joshua.claeys

Reviewed By: joshua.claeys

Subscribers: dane.curbow

Maniphest Tasks: T1453

Differential Revision: https://dev.zeroengine.io/D428