Update tab names on script rename.
Summary:
DocumentEditor listens to resource modified events on the script it is editing and dispatches a TabRenamed event for the TabWidget to recieve and update the tab title.
Moved the assumption that the only tabs being edited are levels and had the editor sending the level renamed event append the "Level: " before the level name.
When updating a resource had the ResourceEvent set the name field to the new name instead of being unused.
Fixes T1513.

Test Plan: Create a zilch script and open it. Rename it. The tab title updates.

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: joshua.claeys, dane.curbow

Maniphest Tasks: T1513

Differential Revision: https://dev.zeroengine.io/D542