Fixed object view focus bug for Shift+W.
Summary:
Changed tree view to react to focus changes events it performs itself for focusing on a selected object.
Fixed bug in ShowRow that did not properly focus on an object when scrolling down to view the row.
Fixes T1550.

Test Plan: Create a object hierarchy with at least 1 child. That child can be given children also. Create enough objects to be able to scroll the hierarchy off screen. Select one of the nested children object and scroll it off screen. Hit Shift+W to select the parent. It will be scrolled into view from above or below. If there are more parents hit shift+W to continue having the parent should it not already be in view (scrolling from below only).

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: joshua.claeys, dane.curbow

Maniphest Tasks: T1550

Differential Revision: https://dev.zeroengine.io/D546