Fixed a bug with SoundEntry EndTime or LoopEndTime being set to zero. + Fixes T2134 -
Summary: The EndTime and LoopEndTime on SoundCue SoundEntries are set to 0 on old projects, which caused them to not play. The field still shows 0 until the sound is changed, but it will now play correctly, and the fields can't be manually set to 0.

Test Plan: Download the project attached to T2134, preview a SoundCue, and verify that it plays.

Reviewers: #zero_engine_developers, dane.curbow

Reviewed By: #zero_engine_developers, dane.curbow

Subscribers: andrea.ellinger

Differential Revision: https://dev.zeroengine.io/D556