Fixed property grid crash when displaying plain 'Resource' type.
Summary:
Added checks on resource manager for the property grids resource view.
When making the base Resource a property it now displays a non-specific list of all resources in the engine.
Adding a new resource opens the regular add window without any specific resource type selected.
Hitting edit resource selects the resource and displays as expected.
Clone/duplicate is disabled in the generic view.
Fixes T1878.-

Test Plan: Create a zilch script with a member of type Resource. Make it a property. Add it to an object and select a resource from the drop down list. Click the create new resource button. Click edit when a resource is selected.

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: joshua.claeys, dane.curbow

Maniphest Tasks: T1878

Differential Revision: https://dev.zeroengine.io/D574