Fixed an issue with copying an object and pasting it into a different space with fewer objects. Fixes T2220 -
Summary: This fixes a problem where an object was trying to be inserted with an index larger than the size of the list.

Test Plan: Copy an object from the end of the object list in a space with lots of objects, and paste it into a space with many fewer objects. It should show up at the end of the list as expected.

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: andrea.ellinger

Differential Revision: https://dev.zeroengine.io/D607