Fixed crash in Find/Replace dialog with unicode characters in search.
Summary: Added check to make sure all characters in a search string are valid to avoid a crash in 'FindTextDialog::GetFindRegex' when 'Regex::Escape' is called.

Test Plan: Search a script with emojis in the search text and make sure a correct warning is displayed and it doesn't cause a crash.

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: joshua.claeys, joshua.shlemmer

Differential Revision: https://dev.zeroengine.io/D626