Fixed a bug causing the TreeView to do unnecessary work.
Summary: When showing a row in the tree view, it was refreshing all values to account for newly expanded rows. However, it was still refreshing if no new rows were even expanded.

Test Plan: Create a Cog. Create 2000+ other Cogs as children of the first. With the children expanding in the object view, clicking on the child Cogs in the scene should not cause the object view to refresh.

Reviewers: #zero_engine_developers, ryan.edgemon

Reviewed By: #zero_engine_developers, ryan.edgemon

Subscribers: joshua.claeys

Differential Revision: https://dev.zeroengine.io/D648