Fixed add window scroll area popping.
Summary: Added check on visible scroll bars ScrollAreaToView back in to restore proper behavior when updating scroll areas. Added UpdateTransform call to the start of MakePositionVisible on the TextEditor to insure that the scroll area has been updated to correctly scroll to the specificed position.

Test Plan: Open the add window and type in the search pop. The list of items to add should not pop back and forth.

Reviewers: #zero_engine_developers, joshua.claeys

Reviewed By: #zero_engine_developers, joshua.claeys

Subscribers: dane.curbow

Differential Revision: https://dev.zeroengine.io/D643