Fixed an issue causing poor performance when selecting a large amount of objects.
Summary: The property grid was enumerating all addable Components just to see if it should show the 'Add Component' button. This was expensive with a large amount of objects (2000+) selected. Now MetaComposition just has a flag that determines whether or not it supports component addition.

Test Plan: Select 2000+ Cogs. It should be faster than it used to. There are still other things causing performance issues, so it may be hard to see the difference.

Reviewers: #zero_engine_developers, ryan.edgemon

Reviewed By: #zero_engine_developers, ryan.edgemon

Subscribers: ryan.edgemon, joshua.claeys

Differential Revision: https://dev.zeroengine.io/D651