When selecting edit resource on an animation with no preview archetype select the resource instead of displaying a warning.
Summary: Fixed bug when selecting to edit an animation and the property grid was not updating to show the previewed archetype and instead was still displaying the levels object. Fixes T1560.

Test Plan: Create an object and add a SimpleAnimation to it. Create a RichAnimation and set it on the SimpleAnimation. In the property grid select edit resource on the animation. It should select it instead of displaying a warning.

Reviewers: #zero_engine_developers, ryan.edgemon

Reviewed By: #zero_engine_developers, ryan.edgemon

Subscribers: dane.curbow

Maniphest Tasks: T1560

Differential Revision: https://dev.zeroengine.io/D716