SoundInstances with an EndTime value will now stop smoothly without clicking. + Fixes T2537 -
Summary: Added a fade when a SoundInstance has an EndTime that's not at the end of the file.

Test Plan: Create a SoundCue with the default sound, and set the Sound's EndTime property to less than its default value. When you preview it there should not be any clicks or other problems.

Reviewers: #zero_engine_developers, nathan.carlson

Reviewed By: #zero_engine_developers, nathan.carlson

Subscribers: andrea.ellinger

Differential Revision: https://dev.zeroengine.io/D726