Fixed issues with usernames containing unicode characters.
Summary:
Fixed bug in find last of when iterating over a string with unicode characters.
Switched all API calls that took a file path to read the contents of the f ile from memory to avoid the issue where an API did not support reading filepaths with unicode characters.
Fixed issue when initilizing the text editor when running the launcher that would result in a null access.
Fixed ImageProcessor/GeometryProcessor to read the wide version of the command line arguements to properly recieve commands containing unicode characters.
Fixes T2536.

Test Plan: Create a username with unicode characters on windows. Open the launcher. Create a new project, make sure the editor exe being launched is the built one from this change. The path will display ???? for unicode characters not in the font face, but will create the project properly in the correct location.

Reviewers: #zero_engine_developers, joshua.davis

Reviewed By: #zero_engine_developers, joshua.davis

Subscribers: joshua.davis, dane.curbow

Maniphest Tasks: T2536

Differential Revision: https://dev.zeroengine.io/D733