Typing a path into a CogPath property will now show a warning if there are multiple objects with the same name. + Fixes T2480 -
Summary: Typing a path directly will now show the same warning that is shown when using the target widget.

Test Plan: In a level with two objects of the same name, verify that using the target widget and typing the path directly both show a warning for a CogPath with an ambiguous path.

Reviewers: #zero_engine_developers, dane.curbow

Reviewed By: #zero_engine_developers, dane.curbow

Subscribers: andrea.ellinger

Differential Revision: https://dev.zeroengine.io/D739