Fix for emitting errors on proxying a component when waiting on plugins to compile. This also prevents proxy errors when loading and scripts don't fully compile. Also prevented "empty" error when plugin is first created. Fixes T879.
Test Plan: Create a plugin (note plugins may currently not compile due to FileStream, but that's not relevant). Close Zero and go to the Level.data and manually add a component with a non-existant name like "Test". Re-open Zero and see that Test is proxied, but errors weren't printed out for proxying.

Reviewers: #zero_engine_developers

Differential Revision: https://dev.zeroengine.io/D749