Fixed issues with command context not getting cleared.
Summary: Added a generic `Context` class that can be used to locally store context, push into global contexts, then remove the local context once done. In the case of ContextMenu, it has a local Context where information can be pushed into (i.e. the parent Cog that was clicked on in the ObjectView for creation commands). When a command is executed from a ContextMenu, the local context is first pushed into the global context. After the command is executed, it will remove the local context from the global context.

Test Plan: Right click on a Cog in the ObjectView and create a Transform underneath it (or any other Cog in that dropdown). Then right click in the ObjectView in open space (not on a Cog) and create another Transform - the newly created Transform should not be parented to the previous parent.

Reviewers: trevor.sundberg, joshua.davis, nathan.carlson

Reviewed By: nathan.carlson

Subscribers: joshua.claeys

Differential Revision: https://dev.zeroengine.io/D725