Fixed a bug caused by modifying an Archetype where there exists a locally added Cog of that Archetype under another Cog of the same Archetype.+
Summary:
Archetype rebuilding was rebuilding both the locally added child Archetype, and its parent. It only needs to rebuild the parent.

It used to walk all the Cogs in the Space which is not in a guaranteed order (moving Cogs around can change the order). Instead, it's now walking the Cogs in hierarchy order. Once it finds a Cog of the Archetype to be rebuilt, it stops recursing that object. Not only does this provide the desired behavior of not getting two objects in the same Hierarchy, it is also an optimization of the old way of doing it.

Test Plan: Create a bunch of Archetype instances in a Level, including locally nesting them (e.g. attach 'House' Archetype to another 'House'). Try upload / reverting a bunch of modifications.

Reviewers: #zero_engine_developers, trevor.sundberg

Reviewed By: #zero_engine_developers, trevor.sundberg

Subscribers: trevor.sundberg, joshua.claeys

Differential Revision: https://dev.zeroengine.io/D436