Added a check when the geometry process fails and goes to remove the failed content file that it was a file is in the local content folder as opposed to a core resource.
Summary: This prevents the odd situation when active development on geometry processor results in a failed import that deletes core resources.

Test Plan: Import a file exported in an unsupported format. It won't be in the content folder after so that it fails to process when you close and open the project again. Rename the bad file to a core resource name (like cube.fbx) and overwrite a core resource with it. Open the engine. When it fails the resource will not be deleted.

Reviewers: #zero_engine_developers, trevor.sundberg

Reviewed By: #zero_engine_developers, trevor.sundberg

Subscribers: trevor.sundberg, dane.curbow

Differential Revision: https://dev.zeroengine.io/D740